requirements.txt
pymongo
youtube_dl
ffmpy
eyed3
Pillow
alive-progress